Mark USB CDC device as not connected when USB connection is lost.
For example, if the cable is pulled out when a terminal is connected to
the device, we won't get a SET_CONTROL_LINE_STATE packet and so the
'connected' field of the driver will remain true. If the usb_poll()
returns false, the device is not configured and so we can assume that
there is no longer an active connection.